move distributed array to utils.
